requirements.txt
flet>=0.10.0
mysql-connector-python>=8.0.0
transformers>=4.30.0
torch>=2.0.0 